Fix client-loop test on python 3.5 and create conftest.py
