+++ jupyter_kernel_mgmt/tests/conftest.py
@@ -1,0 +1,11 @@
+import os
+pjoin = os.path.join
+import pytest
+import sys
+
+@pytest.fixture
+def setup_env(tmpdir, monkeypatch):
+    monkeypatch.setenv('JUPYTER_CONFIG_DIR', pjoin(tmpdir.dirname, 'jupyter'))
+    monkeypatch.setenv('JUPYTER_DATA_DIR', pjoin(tmpdir.dirname, 'jupyter_data'))
+    monkeypatch.setenv('JUPYTER_RUNTIME_DIR', pjoin(tmpdir.dirname, 'jupyter_runtime'))
+    monkeypatch.setenv('IPYTHONDIR', pjoin(tmpdir.dirname, 'ipython'))
