+++ jupyter_kernel_mgmt/tests/utils.py
@@ -5,19 +5,7 @@ import asyncio
 import json
 import os
 pjoin = os.path.join
-import pytest
 import sys
-
-
-skip_win32 = pytest.mark.skipif(sys.platform.startswith('win'), reason="Windows")
-
-
-@pytest.fixture
-def setup_env(tmpdir, monkeypatch):
-    monkeypatch.setenv('JUPYTER_CONFIG_DIR', pjoin(tmpdir.dirname, 'jupyter'))
-    monkeypatch.setenv('JUPYTER_DATA_DIR', pjoin(tmpdir.dirname, 'jupyter_data'))
-    monkeypatch.setenv('JUPYTER_RUNTIME_DIR', pjoin(tmpdir.dirname, 'jupyter_runtime'))
-    monkeypatch.setenv('IPYTHONDIR', pjoin(tmpdir.dirname, 'ipython'))
 
 
 def execute(code='', kc=None, **kwargs):
